improve received events enum implemntation

+++ models/player.go
@@ -1,8 +1,8 @@
 package models
 
 type Position struct {
-	X float32 `json:"x"`
-	Y float32 `json:"y"`
+	X float64 `json:"x"`
+	Y float64 `json:"y"`
 }
 type Player struct {
 	ID       string   `json:"id"`
